[FIX] loyalty_criteria_multi_product: fix test
The good use of this criteria should really be based on the proper
configuration of a rule, when there is more than one rule, as soon as
one is fulfilled, all of them are fulfilled, functionally it can be
beneficial but in the execution of the tests it can interfere with other
promotions and not be carrying out the proper testing of the module.
For this reason, the configuration of the promotion is reduced to a
single rule with multi-product criteria.

TT48566